+++ project/src/main/java/lu/uni/adtool/ui/canvas/AbstractTreeCanvas.java
@@ -192,6 +192,16 @@ public abstract class AbstractTreeCanvas extends JPanel
     }
   }
 
+  public void setNodeLayout(NodeLayout layout) {
+    if (layout != this.nodeLayout) {
+      this.nodeLayout = layout;
+      this.recalculateLayout();
+      this.repaint();
+    }
+  }
+  public NodeLayout getNodeLayout() {
+    return this.nodeLayout;
+  }
   public abstract void setScrollPane(JScrollPane pane);
 
   public abstract void repaintAll();
@@ -584,7 +594,7 @@ public abstract class AbstractTreeCanvas extends JPanel
    *
    */
   protected void recalculateLayout() {
-    Debug.log("tree " + tree);
+//     Debug.log("tree " + tree);
     this.sizeX = 0;
     this.sizeY = 0;
     bufferedLayout = null;
@@ -603,7 +613,7 @@ public abstract class AbstractTreeCanvas extends JPanel
        sizeX = Math.max(sizeX, rect.getMaxX());
        sizeY = Math.max(sizeY, rect.getMaxY());
     }
-    if(this.nodeLayout == NodeLayout.RADIAL) {
+    if (this.nodeLayout == NodeLayout.RADIAL) {
       Node root = this.tree.getRoot(false);
       Rectangle2D.Double rect = bufferedLayout.get(root);
       radialX = rect.getCenterX();
@@ -614,8 +624,8 @@ public abstract class AbstractTreeCanvas extends JPanel
       double maxY = Double.MIN_VALUE;
       for (Node node: bufferedLayout.keySet()) {
         rect = bufferedLayout.get(node);
-        double angle = ((rect.getCenterX() - radialX)/sizeX)*2*Math.PI;
-        double r = rect.getCenterY() - radialY;
+        double angle = -((rect.getCenterX() - radialX)/sizeX)*2*Math.PI;
+        double r = Options.canv_radialRadiusScale * (rect.getCenterY() - radialY);
 //         Debug.log("r:" + r + " angle:"+angle);
 //         Debug.log("getX:" + rect.getX() + " getY:"+rect.getY() + " w:"+ rect.getWidth()+ " h:"+ rect.getHeight()+ " name:" + node.getName());
         rect.setRect(r*Math.cos(angle) + radialX - rect.getWidth()/2, r*Math.sin(angle) + radialY- rect.getHeight()/2, rect.getWidth(), rect.getHeight());
@@ -1251,7 +1261,6 @@ public abstract class AbstractTreeCanvas extends JPanel
         }
         g2.setStroke(basicStroke);
 
-//         if (
         if (drawArc) {
           double tangens1 = (b2.getCenterY() - (double) y1) / (minX - (double) x1);
           double tangens2 = (b2.getCenterY() - (double) y1) / (maxX - (double) x1 - 1);
@@ -1287,22 +1296,21 @@ public abstract class AbstractTreeCanvas extends JPanel
         }
       }
       else { //RADIAL layout
-//         g2.drawLine((int) x1, (int) y1, (int) radialX, (int) b2.getCenterY());
-//         rect.setRect(r*Math.cos(angle) + radialX - rect.getWidth()/2, r*Math.sin(angle) + radialY- rect.getHeight()/2, rect.getWidth(), rect.getHeight());
         if (children.size() > 0) {
           Rectangle2D.Double bParen = bufferedLayout.get(parent);
           double tempX = bParen.getCenterX() - radialX;
           double tempY = bParen.getCenterY() - radialY;
           double parenR = Math.sqrt((tempX*tempX + tempY*tempY));
-          double minAngle = Double.MAX_VALUE;
+          double firstAngle = Double.MAX_VALUE;
+          double counterAngle = Double.MIN_VALUE;
           double lastAngle = Double.MIN_VALUE;
-          double maxAngle = Double.MIN_VALUE;
           double r = -1;
           double halfR = -1;
           double meanX = 0;
           double meanY = 0;
 
           int i = 1;
+          //draw lines
           for (Node child : children) {
             Rectangle2D.Double b2 = bufferedLayout.get(child);
             tempX = b2.getCenterX() - radialX;
@@ -1310,23 +1318,20 @@ public abstract class AbstractTreeCanvas extends JPanel
             meanX = meanX + tempX;
             meanY = meanY + tempY;
             if (r < 0) {
-              //assumnig r is the same for all children
+              //assumnig radius r is the same for all children
               r = Math.sqrt(tempX*tempX + tempY*tempY);
               halfR = (parenR + r)/2;
             }
-//             double angle = Math.signum(tempY)*Math.acos(Math.max(Math.min(tempX/r, 1), -1));
             double angle = (Math.atan2(-tempY, tempX) + 2 * Math.PI) % (2*Math.PI);
-//             Debug.log("child no"+Integer.toString(i) + " r=" +Double.toString(r)+ " tempaX=" +Double.toString(tempX)+ " angle=" +Double.toString(angle));
-            Debug.log("angle="+Double.toString(Math.toDegrees(angle)) + " i:"+Integer.toString(i) + " child.size():"+Integer.toString( children.size())+ " label:"+child.getName());
-
+//             Debug.log("angle="+Double.toString(Math.toDegrees(angle)) + " i:"+Integer.toString(i) + " child.size():"+Integer.toString( children.size())+ " label:"+child.getName());
             if (i == 1) {
-              minAngle = angle;
+              firstAngle = angle;
             }
             if (i == children.size()) {
-              lastAngle = angle;
+              counterAngle = angle;
             }
             if (i == noChildren ) {
-              maxAngle = angle;
+              lastAngle = angle;
             }
             if (i > noChildren) {
               g2.setStroke(counterStroke);
@@ -1337,42 +1342,55 @@ public abstract class AbstractTreeCanvas extends JPanel
             g2.drawLine((int)(r*Math.cos(angle) + radialX), (int)(-r*Math.sin(angle) + radialY), (int)(halfR*Math.cos(angle) + radialX), (int)(-halfR*Math.sin(angle) + radialY));
             i = i + 1;
           }
-          if (maxAngle == Double.MIN_VALUE) {
-            maxAngle = (minAngle + lastAngle)/2;
+          while (counterAngle < firstAngle) {
+            counterAngle += 2*Math.PI;
           }
-          if (noChildren==0) {
-            g2.setStroke(counterStroke);
+          if (lastAngle == firstAngle) {
+            lastAngle = (firstAngle + counterAngle)/2;
           }
-//           double midAngle = (minAngle - maxAngle)/2;
-          meanX = meanX/i;
-          meanY = meanY/i;
-          double midAngle = (Math.atan2(-meanY, meanX) + 2 * Math.PI) % (2*Math.PI);
-          Arc2D arc = new Arc2D.Double(radialX - halfR,
-                                       radialY - halfR,
-                                       2*halfR, 2*halfR,
-                                       Math.toDegrees(minAngle),
-                                       -Math.toDegrees((minAngle - maxAngle + 2*Math.PI)%(2*Math.PI)),
-                                       Arc2D.OPEN);
-//           Arc2D arc = new Arc2D.Double(radialX - halfR,
-//                                        radialY - halfR,
-//                                        2*halfR, 2*halfR,
-//                                        Math.toDegrees(minAngle),
-//                                        90,
-//                                        Arc2D.OPEN);
-
-          Debug.log("minAngle="+Double.toString(Math.toDegrees(minAngle)) +  "maxAngle="+Double.toString(Math.toDegrees(maxAngle)));
-          Debug.log("lastAngle="+Double.toString(Math.toDegrees(lastAngle)) +  "midAngle="+Double.toString(Math.toDegrees(midAngle)));
-          g2.draw(arc);
-
+          if (noChildren > 0) {
+            g2.setStroke(basicStroke);
+          }
+          double midAngle = (firstAngle + counterAngle)/2;
           g2.drawLine((int)(parenR*Math.cos(midAngle) + radialX), (int)(-parenR*Math.sin(midAngle) + radialY), (int)(halfR*Math.cos(midAngle) + radialX), (int)(-halfR*Math.sin(midAngle) + radialY));
 
-//           Rectangle2D.Double bFirst = bufferedLayout.get(children.get(0));
-//           Rectangle2D.Double bLast = bufferedLayout.get(children.get(noChildren - 1));
-//           double firstX = bFirst.getCenterX() - radialX;
-//           double firstY = bFirst.getCenterY() - radialY;
-//           double firstR = Math.sqrt((firstX*firstX + firstY*firstY));
-//           double firstAngl = Math.signum((tempY-radialY)) * Math.acos(firstX/firstR);
-//           double lastAngl = Math.signum((tempY-radialY)) * Math.acos(bLast.getCenterX() - radialX/firstR);
+          meanX = meanX/i;
+          meanY = meanY/i;
+          //draw arc for non-countered children
+          if (noChildren > 0) {
+            Arc2D arc = new Arc2D.Double(radialX - halfR,
+                                       radialY - halfR,
+                                       2*halfR, 2*halfR,
+                                       Math.toDegrees(firstAngle),
+                                       Math.toDegrees((lastAngle - firstAngle + 2*Math.PI)%(2*Math.PI)),
+                                       Arc2D.OPEN);
+            g2.draw(arc);
+          }
+          //draw arc for OR operator
+          if (drawArc) {
+            Arc2D arc = new Arc2D.Double(radialX - halfR - Options.canv_ArcPadding/2,
+                                       radialY - halfR - Options.canv_ArcPadding/2,
+                                       2*halfR + Options.canv_ArcPadding, 2*halfR + Options.canv_ArcPadding,
+                                       Math.toDegrees(firstAngle),
+                                       Math.toDegrees((lastAngle - firstAngle + 2*Math.PI)%(2*Math.PI)),
+                                       Arc2D.OPEN);
+
+            g2.draw(arc);
+          }
+          //draw arc for countered child
+          if (noChildren < children.size() && noChildren > 0) {
+            g2.setStroke(counterStroke);
+            Arc2D arc = new Arc2D.Double(radialX - halfR,
+                                       radialY - halfR,
+                                       2*halfR, 2*halfR,
+                                       Math.toDegrees(lastAngle),
+                                       Math.toDegrees((counterAngle - lastAngle + 2*Math.PI)%(2*Math.PI)),
+                                       Arc2D.OPEN);
+            g2.draw(arc);
+          }
+//           Debug.log("firstAngle="+Double.toString(Math.toDegrees(firstAngle)) +  "lastAngle="+Double.toString(Math.toDegrees(lastAngle)));
+//           Debug.log("counterAngle="+Double.toString(Math.toDegrees(counterAngle)) +  "midAngle="+Double.toString(Math.toDegrees(midAngle))+ " noChildren:"+ Integer.toString(noChildren));
+
         }
       }
       for (Node child : children) {
@@ -1561,7 +1579,7 @@ public abstract class AbstractTreeCanvas extends JPanel
   public static enum NodeLayout{
     NORMAL, RADIAL
   }
-  protected NodeLayout                    nodeLayout = NodeLayout.RADIAL;
+  protected NodeLayout                    nodeLayout = Options.canv_defaultLayout;
   /**
    * Center of root node when using RADIAL display layout - x component
    */